https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved jmapClientProvider to esn.inbox.libs
