+++ src/esn.inbox.libs/app/services/jmap-client-provider/jmap-client-provider.run.js
@@ -1,0 +1,8 @@
+'use strict';
+
+require('./jmap-client-provider.js');
+
+angular.module('esn.inbox.libs')
+  .run(function(jmapClientProvider) {
+    return jmapClientProvider.get();
+  });
