+++ src/linagora.esn.unifiedinbox/app/services.js
@@ -9,7 +9,7 @@ require('../../esn.inbox.libs/app/services/hook/email-sending-hook.service.js');
 require('../../esn.inbox.libs/app/services/email-body/email-body.js');
 require('./services/attachment-jmap/attachment-jmap.constants.js');
 require('./services/email-resolver/inbox-email-resolver.service.js');
-require('./services/jmap-client-provider/jmap-client-provider.js');
+require('../../esn.inbox.libs/app/services/jmap-client-provider/jmap-client-provider.js');
 
 (function (angular) {
   'use strict';
