--- src/linagora.esn.unifiedinbox/app/services/jmap-client-provider/jmap-client-provider.run.js
@@ -1,10 +1,0 @@
-require('../jmap-client-provider/jmap-client-provider.js');
-
-(function(angular) {
-  'use strict';
-
-  angular.module('linagora.esn.unifiedinbox')
-    .run(function(jmapClientProvider) {
-      return jmapClientProvider.get();
-    });
-})(angular);
